requirements.txt
streamlit>=1.32.0
google-cloud-bigquery>=3.17.0
google-auth>=2.27.0
db-dtypes>=1.2.0
pandas>=2.0.0
pydeck>=0.8.0
